#feature/FPS Test Supplier View Sonar fix

+++ src/Apha.FPSApps/Apha.FPSApps.Application/Dtos/FPS/TestSupplierViewDto.cs
@@ -5,7 +5,7 @@ namespace Apha.FPSApps.Application.Dtos.FPS
         public string TestCode { get; set; } = null!;
         public string Buyer { get; set; } = null!;
         public string? ProjectManager { get; set; }
-        public double? NoRequired { get; set; }
+        public int? NoRequired { get; set; }
         public decimal? UnitPrice { get; set; }
         public decimal? TestCost { get; set; }
         public string? ProjectStatus { get; set; }
